+++ apps/prairielearn/src/lib/grading.sql
@@ -9,14 +9,6 @@ FROM
 WHERE
   s.id = $submission_id;
 
--- BLOCK select_assessment_instance_id_for_instance_question
-SELECT
-  iq.assessment_instance_id
-FROM
-  instance_questions AS iq
-WHERE
-  iq.id = $instance_question_id;
-
 -- BLOCK select_workspace_id
 SELECT
   w.id AS workspace_id
