+++ apps/prairielearn/src/lib/grading.ts
@@ -10,13 +10,7 @@ import { IdSchema, IntervalSchema } from '@prairielearn/zod';
 import { updateCourseInstanceUsagesForSubmission } from '../models/course-instance-usages.js';
 import { insertGradingJob, updateGradingJobAfterGrading } from '../models/grading-job.js';
 import { computeNextAllowedGradingTimeMs } from '../models/instance-question.js';
-import {
-  type ResolvedSharedStateObject,
-  extractSharedStateDefaultsForObjects,
-  readSharedStateValuesForAssessmentInstance,
-  selectSharedStateObjectsForQuestion,
-  writeSharedStateValuesForAssessmentInstance,
-} from '../models/shared-state-value.js';
+import { writeSharedStateValuesForAssessmentInstance } from '../models/shared-state-value.js';
 import { lockVariant } from '../models/variant.js';
 import * as questionServers from '../question-servers/index.js';
 
@@ -27,7 +21,6 @@ import {
   InstanceQuestionSchema,
   type Question,
   QuestionSchema,
-  type SharedStateObjectValue,
   type Submission,
   SubmissionSchema,
   type Variant,
@@ -39,98 +32,10 @@ import { writeCourseIssues } from './issues.js';
 import * as ltiOutcomes from './ltiOutcomes.js';
 import { updateInstanceQuestionStats } from './question-points.js';
 import { getQuestionCourse } from './question-variant.js';
-import {
-  diffSharedStateObjectValue,
-  validateSharedStateObjectValueForWrite,
-} from './shared-state.js';
+import { resolveSharedStateForPhase, validateSharedStatePatch } from './shared-state-grading.js';
 import * as workspaceHelper from './workspace.js';
 
 const sql = sqldb.loadSqlEquiv(import.meta.url);
-
-interface SharedStateResolution {
-  objects: Record<string, ResolvedSharedStateObject>;
-  assessment_instance_id: string | null;
-  before: Record<string, SharedStateObjectValue>;
-}
-
-/**
- * Resolves the shared-state objects a question can access and reads their
- * current, live value for the assessment instance a variant belongs to. A
- * variant with no assessment instance (e.g. an instructor preview) has
- * nothing to read live, but `data["shared_state"]` is still populated with
- * each object's schema defaults — mirroring how `question-variant.ts` seeds
- * `generate`/`prepare` for the same case — rather than an empty object.
- */
-async function resolveSharedStateForGrading({
-  question,
-  question_course,
-  instance_question_id,
-}: {
-  question: Question;
-  question_course: Course;
-  instance_question_id: string | null;
-}): Promise<SharedStateResolution> {
-  const objects = await selectSharedStateObjectsForQuestion({
-    course_id: question_course.id,
-    question,
-  });
-
-  if (Object.keys(objects).length === 0) {
-    return { objects, assessment_instance_id: null, before: {} };
-  }
-
-  const assessment_instance_id =
-    instance_question_id == null
-      ? null
-      : await sqldb.queryOptionalScalar(
-          sql.select_assessment_instance_id_for_instance_question,
-          { instance_question_id },
-          IdSchema,
-        );
-  if (assessment_instance_id == null) {
-    return {
-      objects,
-      assessment_instance_id: null,
-      before: extractSharedStateDefaultsForObjects(objects),
-    };
-  }
-
-  const before = await readSharedStateValuesForAssessmentInstance({
-    assessment_instance_id,
-    objects,
-  });
-  return { objects, assessment_instance_id, before };
-}
-
-/**
- * Validates the property-level patch a question phase produced against each
- * object's schema, merged onto the pre-phase snapshot. This is a cheap
- * pre-check so an invalid write becomes a fatal course issue (and the
- * question is marked broken) before the submission/grading result is
- * persisted, rather than only being caught later when the patch is actually
- * written back to the database.
- */
-function validateSharedStatePatch({
-  objects,
-  before,
-  after,
-}: {
-  objects: Record<string, ResolvedSharedStateObject>;
-  before: Partial<Record<string, SharedStateObjectValue>>;
-  after: Partial<Record<string, SharedStateObjectValue>>;
-}): string[] {
-  const issues: string[] = [];
-  for (const [name, object] of Object.entries(objects)) {
-    const patch = diffSharedStateObjectValue(before[name] ?? {}, after[name] ?? before[name] ?? {});
-    if (Object.keys(patch).length === 0) continue;
-    const merged = { ...before[name], ...patch };
-    const errors = validateSharedStateObjectValueForWrite(merged, object.properties);
-    if (errors.length > 0) {
-      issues.push(`Shared-state object "${name}": ${errors.join('; ')}`);
-    }
-  }
-  return issues;
-}
 
 const VariantDataSchema = z.object({
   grading_method: QuestionSchema.shape.grading_method,
@@ -339,7 +244,7 @@ export async function saveSubmission(
 
   const questionModule = questionServers.getModule(question.type);
   const question_course = await getQuestionCourse(question, variant_course);
-  const sharedState = await resolveSharedStateForGrading({
+  const sharedState = await resolveSharedStateForPhase({
     question,
     question_course,
     instance_question_id: variant.instance_question_id,
@@ -545,7 +450,7 @@ export async function gradeVariant({
     // For Internal grading we call the grading code. For Manual grading, if the question
     // reached this point, it has auto points, so it should be treated like Internal.
     const questionModule = questionServers.getModule(question.type);
-    const sharedState = await resolveSharedStateForGrading({
+    const sharedState = await resolveSharedStateForPhase({
       question,
       question_course,
       instance_question_id: variant.instance_question_id,
